requirements.txt
textblob
nltk
matplotlib
python-Levenshtein
customtkinter